+++ code/.coveragerc
@@ -1,0 +1,6 @@
+[run]
+relative_files = True
+include =
+  network_manger/*
+  tests/*
+branch = True
